+++ crates/common/src/lib.rs
@@ -4,6 +4,7 @@ pub mod auth_provider;
 pub mod db;
 pub mod db_analytics;
 pub mod db_auth;
+pub mod live_event;
 pub mod model;
 
 use anyhow::Result;